refactor: split initial data from crud
